requirements.txt
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.1.0
asgiref==3.8.1
dj-database-url==2.2.0
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
beniget==0.4.1
bleach==4.1.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.5.0
certifi==2020.6.20
cffi==1.17.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.2.1
cryptography==43.0.0
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-crontab==0.7.1
django-heroku==0.3.1
djangorestframework==3.15.2
duplicity==0.8.21
entrypoints==0.4
fasteners==0.14.1
fonttools==4.53.1
fs==2.4.12
future==0.18.2
gast==0.5.2
geographiclib==2.0
geopy==2.4.1
google-api-core==2.20.0
google-auth==2.35.0
google-cloud==0.34.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
gunicorn==23.0.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.4.2
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.3
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
notebook==6.4.8
numpy==2.1.0
oauthlib==3.2.0
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.4.0
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.28
proto-plus==1.24.0
protobuf==5.28.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pypdfium2==4.30.0
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==1.0.1
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
reportlab==3.6.8
requests==2.25.1
rsa==4.9
schedule==1.2.2
scikit-learn==1.5.1
scipy==1.14.0
SecretStorage==3.3.1
Send2Trash==1.8.1b0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.0
sympy==1.9
systemd-python==234
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.5.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.12.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.7.0
widgetsnbextension==2.0.0
xdg==5
xkit==0.0.0
zipp==1.0.0
